Move logger to namespace in separate file
