+++ src/logger.cpp
@@ -1,0 +1,20 @@
+#include <boost/date_time.hpp>
+#include <nlohmann/json.hpp>
+
+using nlohmann::json;
+
+namespace logger {
+
+  void log(std::string event, json j) {
+
+    boost::posix_time::ptime timestamp { boost::posix_time::microsec_clock::universal_time() };
+
+    j["timestamp"] = to_iso_extended_string(timestamp);
+    j["event"] = event;
+
+    std::cout << j.dump() << std::endl;
+
+  }
+
+}
+
